RenderInterface Updates and Added testing code (Draws a simple triangle using index and vertex buffers)
UPDATE:
- CMakeLists.txt: Debug and Release builds now different
- Profiler.h: Updated release marco to not break the build
- GameEngine.h/.cpp: Added testing code for render interface

- VulkanCommandBuffer.h/.cpp: Added SetWaitFence/ResetWaitFence functions
- VulkanSwapChain: added a method for aquiring the next image index
- VulkanPipeline.h: fixed multistatecreateinfo bug that was setting the sample mask to 0 always instead of nullptr when bAlphaToCoverage is not enabled
- VulkanShader.cpp: uses the entry point provided by the shader configuration passed in

- CommandBuffer.h: Added a getter for the waitfence in use by the command buffer
- CommandQueue.h: added a way to ResetWaitFence
- RenderPassGenerics.h added ETextureLayout (specifies the texture layout)
- SwapChain.h: added a method for acquiring the next swapchain image index

+++ Source/Runtime/Graphics/PipelineLayout.h
@@ -84,7 +84,7 @@ public:
     */
     inline virtual uint32 GetNumBinding() const
     {
-        return Bindings.size();
+        return (uint32)Bindings.size();
     }
 
     /**
